Move camera poses to vision constants and add vision multiplier constants
